feat: extend mute/fade functionality for AUX and FX buses
Setting fader levels and muting channels was only possible for the master bus.
This commit extends the functionality: Fader levels and mute can now be set on AUX and FX buses as well.
For this, the action has two new options now: bus type (master/aux/fx) and the bus number. The bus number must only be provided if the bus is not master.

This commit also brings some minor code formatting changes.
Resolve #4